+++ lib_xcore_math/src/arch/ref/vect_convolve.c
@@ -1,13 +1,13 @@
-// Copyright 2020-2022 XMOS LIMITED.
+// Copyright 2020-2023 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
 
 #include <stdint.h>
 #include <stdio.h>
 
 #include "xmath/xmath.h"
-#include "../../vect/vpu_helper.h"
+#include "vpu_helper.h"
 #include "xmath/xs3/vpu_scalar_ops.h"
-#include "../../vect/vpu_const_vects.h"
+#include "vpu_const_vects.h"
 
 
 
@@ -21,14 +21,14 @@ headroom_t vect_s32_convolve_valid(
   const unsigned P = b_length >> 1;
   const unsigned y_length = x_length - 2*P;
 
-  for(int k = 0; k < y_length; k++){
+  for(unsigned k = 0; k < y_length; k++){
 
     vpu_int32_acc_t acc = 0;
-    for(int t = 0; t < b_length; t++){
+    for(unsigned t = 0; t < b_length; t++){
       acc = vlmacc32(acc, x[k+t], b_q30[t]);
     }
 
-    y[k] = acc;
+    y[k] = (int32_t) acc;
   }
 
   return vect_s32_headroom(y, y_length);
