+++ test/shared/floating_fft/floating_fft_float.c
@@ -16,7 +16,7 @@ void flt_bit_reverse_indexes_float(
     const unsigned length)
 {
     size_t logn = u32_ceil_log2(length);
-    for(int i = 0; i < length; i++){
+    for(unsigned int i = 0; i < length; i++){
         unsigned rev = flt_bitrev(i, logn);
         if(rev < i) continue;
 
@@ -28,7 +28,7 @@ void flt_bit_reverse_indexes_float(
 }
 
 void flt_make_sine_table_float(
-    float sine_lut[], 
+    float sine_lut[],
     const unsigned N)
 {
     for(unsigned i = 0; i < (N/4) + 1; i++){
@@ -148,14 +148,14 @@ void flt_fft_inverse_float (
     unsigned log2_N = 31-cls(N);
 
     for(unsigned i=0;i<N;i++){
-        pts[i].re = ldexpf(pts[i].re, -log2_N);
-        pts[i].im = ldexpf(pts[i].im, -log2_N);
+        pts[i].re = ldexpf(pts[i].re, -(int)log2_N);
+        pts[i].im = ldexpf(pts[i].im, -(int)log2_N);
     }
 }
 
 
-void flt_fft_split_spectrum_float( 
-    complex_float_t pts[], 
+void flt_fft_split_spectrum_float(
+    complex_float_t pts[],
     const unsigned N)
 {
 
@@ -189,7 +189,7 @@ void flt_fft_split_spectrum_float(
 
 
 void flt_fft_merge_spectra_float(
-    complex_float_t pts[], 
+    complex_float_t pts[],
     const unsigned N)
 {
     for(uint32_t i=1; i < (N/4); i++){
